pyproject.toml
[project]
name = "ammoseekr"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.1.0",
    "pydantic>=2.10.6",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pulumi>=3.0.0,<4.0.0",
    "pulumi-gcp>=8.0.0,<9.0.0",
    "pytest>=8.3.5",
    "ruff>=0.11.0",
]

[tool.ruff.lint]
select = ["F", "B", "E", "SIM", "I"]
ignore = ["E501"]
fixable = ["ALL"]
